+++ src/core/dpki.js
@@ -21,7 +21,7 @@
  */
 
 import * as tendermintNdid from '../tendermint/ndid';
-import { validateKeyType } from './common';
+import { validateKey } from './common';
 import logger from '../logger';
 import { callbackToClient } from '../utils/callback';
 import { getErrorObjectForClient } from '../error/helpers';
@@ -39,11 +39,11 @@ export async function updateNode(
 ) {
   // Validate public keys
   if (public_key != null) {
-    validateKeyType(public_key, public_key_type);
+    validateKey(public_key, public_key_type);
   }
 
   if (master_public_key != null) {
-    validateKeyType(master_public_key, master_public_key_type);
+    validateKey(master_public_key, master_public_key_type);
   }
 
   if (synchronous) {