+++ src/core/ndid.js
@@ -21,7 +21,7 @@
  */
 
 import * as tendermint from '../tendermint';
-import { validateKeyType } from './common';
+import { validateKey } from './common';
 import logger from '../logger';
 import * as utils from '../utils';
 import { callbackToClient } from '../utils/callback';
@@ -45,11 +45,11 @@ export async function initNDID({
   }
   try {
     if (public_key != null) {
-      validateKeyType(public_key, public_key_type);
+      validateKey(public_key, public_key_type);
     }
 
     if (master_public_key != null) {
-      validateKeyType(master_public_key, master_public_key_type);
+      validateKey(master_public_key, master_public_key_type);
     }
 
     await tendermint.transact(
@@ -127,11 +127,11 @@ export async function reduceNodeToken(data) {
 export async function registerNode(data, { synchronous = false } = {}) {
   try {
     if (data.public_key != null) {
-      validateKeyType(data.public_key, data.public_key_type);
+      validateKey(data.public_key, data.public_key_type);
     }
 
     if (data.master_public_key != null) {
-      validateKeyType(data.master_public_key, data.master_public_key_type);
+      validateKey(data.master_public_key, data.master_public_key_type);
     }
 
     if (synchronous) {